app/agent/requirements.txt
livekit
livekit-agents
livekit-plugins-silero
python-dotenv
aiohttp

# Hugging Face and ML dependencies (for STT/TTS/LLM)
transformers
torch
torchaudio
accelerate
sentencepiece
protobuf

# For TTS (Text-to-Speech) - Using transformers + speecht5
soundfile
scipy
librosa
datasets

# For STT (Speech-to-Text) - Whisper
openai-whisper
ffmpeg-python

# For LLM inference
bitsandbytes

